PP-6072 Allow searching for gateway account by more parameters
Previously, could just search by a list of gateway account ids. Now
can provide an additional query parameter `moto_enabled` to filter
gateway accounts by whether or not they allow moto payments or not.

Use a bean for the query parameters for the get gateway accounts
endpoint and a new search method that was added to the DAO previously.
Clean up the old methods that are no longer needed now.